pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-trading"
version = "2.1.0"
description = "Market Architect Pro - Trading Platform Dashboard"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "streamlit",
    "alpaca-py",
    "pandas",
    "numpy",
    "yfinance",
    "pandas-ta",
    "scikit-learn",
    "plotly",
    "requests",
    "python-dotenv",
    "textblob",
    "python-docx",
]
requires-python = ">=3.8"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "views*"]
exclude = ["venv*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]